long double -> double globally; strip L suffix from float literals
Cell normalization keeps probabilities in [0,1]; 64-bit double has plenty
of precision for 47-214 hap panels. Stripping the L from 1.0L/0.0L is
critical: without it those literals stay long double and force every
arithmetic expression to promote back to 80-bit x87, undoing the speedup.

Existing 'set to uniform' underflow handler unchanged; benchmarks show
zero genotype divergence vs v4.2.1 on chr22 (331,933/331,933 = 100%)
and tiny region (6,605/6,605 = 100%).

+++ src/copynumber.hpp
@@ -13,13 +13,13 @@ public:
 	@param cn_1 probability of copy number 1
 	@param cn_2 probability of copy number 2
 	 **/
-	CopyNumber(long double cn_0, long double cn_1, long double cn_2);
-	CopyNumber(long double cn_0, long double cn_1, long double cn_2, long double regularization_const);
+	CopyNumber(double cn_0, double cn_1, double cn_2);
+	CopyNumber(double cn_0, double cn_1, double cn_2, double regularization_const);
 	/** get probability of copy number cn **/
-	long double get_probability_of(int cn) const;
+	double get_probability_of(int cn) const;
 	bool operator==(const CopyNumber &other) const;
 	bool operator!=(const CopyNumber &other) const;
 private:
-	std::vector<long double> probabilities;
+	std::vector<double> probabilities;
 };
 #endif // COPYNUMBER_H
